requirements.txt
mysql-connector-python==8.0.33
pandas==2.0.3
pyarrow==12.0.0
python-dotenv==1.0.0
pytest==7.4.0